#821 [REFACTOR] 폴더 구조 정리 - feature/search
