--- src/components/Search/index.js
@@ -1,1 +1,0 @@
-export { default as Search } from './Search';
