+++ src/constants/index.js
@@ -1,2 +1,1 @@
-export * from './search.js';
 export * from './seconds.js';
