+++ src/feature/search/component/index.js
@@ -1,0 +1,1 @@
+export { default as Search } from './Search/Search';
