+++ src/feature/search/constant/index.js
@@ -1,0 +1,1 @@
+export * from './search.js';
