+++ src/pages/BoardPage/BoardPage.jsx
@@ -4,7 +4,7 @@ import { Header, Icon } from '@/shared/component';
 import { BOARD_MENUS } from '@/shared/constant';
 
 import { BoardBar } from '@/feature/board/component';
-import { Search } from '@/components';
+import { Search } from '@/feature/search/component';
 
 import styles from './BoardPage.module.css';
