+++ src/pages/SearchPage/components/SearchResults/SearchResults.jsx
@@ -1,7 +1,5 @@
 import { useLocation, Link } from 'react-router-dom';
 
-import { useSearch } from '@/hooks';
-
 import { FetchLoading, List, PullToRefresh } from '@/shared/component';
 import {
   getBoardTitleToTextId,
@@ -11,6 +9,7 @@ import {
 import { BOARDS } from '@/shared/constant';
 
 import { PostBar } from '@/feature/board/component';
+import { useSearch } from '@/feature/search/hook';
 
 import styles from './SearchResults.module.css';
